修改签到列表 修改环信登录 修改webview加载统计

+++ eanfang/src/main/java/com/eanfang/model/SignListBean.java
@@ -278,7 +278,7 @@ public class SignListBean implements Serializable {
             private int sortNum;
             private Long topCompanyId;
             private String updateTime;
-            private int updateUser;
+            private Long updateUser;
 
             public Long getCompanyId() {
                 return companyId;
@@ -368,11 +368,11 @@ public class SignListBean implements Serializable {
                 this.updateTime = updateTime;
             }
 
-            public int getUpdateUser() {
+            public Long getUpdateUser() {
                 return updateUser;
             }
 
-            public void setUpdateUser(int updateUser) {
+            public void setUpdateUser(Long updateUser) {
                 this.updateUser = updateUser;
             }
         }
